wallet: extract CreateImportedAccount shared workflow
